requirements.txt
# Dependencies
pandas~=2.2.2
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1
plotly~=5.22.0
psycopg2-binary~=2.9.9
streamlit~=1.45.0
xlrd ~= 2.0.1